Rename DependencyRule -> RawDependencyRule
