--- src/types/DependencyRule.ts
@@ -1,4 +1,0 @@
-import FullDependencyRule from './FullDependencyRule';
-
-type DependencyRule = FullDependencyRule | string;
-export default DependencyRule;
