+++ src/types/rawConfig/RawDependencyRule.ts
@@ -1,0 +1,4 @@
+import FullDependencyRule from '../FullDependencyRule';
+
+type RawDependencyRule = FullDependencyRule | string;
+export default RawDependencyRule;
